% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Matern.NS.Cov.R
\name{Anisotropic.Matern.Nonstationary.Cov}
\alias{Anisotropic.Matern.Nonstationary.Cov}
\title{Define a nonstationary covariance matrix based on a generalization of the Matern with geometric anisotropy}
\usage{
Anisotropic.Matern.Nonstationary.Cov(
  x,
  y,
  Lx,
  Ly,
  ang,
  nu,
  sigma = 1,
  verbose = FALSE
)
}
\arguments{
\item{x}{x-coordinate of locations on a regular rectangular grid}

\item{y}{y-coordinate of locations on a regular rectangular grid}

\item{Lx}{range in x direction (geometric anistotropy)}

\item{Ly}{range in x direction (geometric anistotropy)}

\item{ang}{angle of rotation (geometric anistotropy)}

\item{nu}{smoothness parameter}

\item{sigma}{standard deviation parameter}

\item{verbose}{whether to print progress or not}
}
\value{
a covariance matrix
}
\description{
Encode (local) spatially varying parameters (ranges Lx and Ly, smoothness nu,
rotation angle, variance sigma) into a global nonstationarity covariance with
local geometric anisotropy. See Paciorek (2003) and Stein (2005).
}
