% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.ns.R
\name{fit.fivepoint.J.s}
\alias{fit.fivepoint.J.s}
\title{Fit nonstationary GMRF precision matrix from data}
\usage{
fit.fivepoint.J.s(xx, yy, ZZ, symm, init = NULL)
}
\arguments{
\item{xx}{x-coordinates of data}

\item{yy}{y-coordinates of data}

\item{ZZ}{response variable of data at locations}

\item{symm}{if TRUE, defines H using the symmetric square root, otherwise uses the Cholesky}

\item{init}{initial values in optimization}
}
\value{
results from optim
}
\description{
Fit nonstationary GMRF precision matrix from data
}
