% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Matern.logLik.aniso.R
\name{fit.Matern.aniso}
\alias{fit.Matern.aniso}
\title{Wrapper function for fitting an anisotropic Matern GP}
\usage{
fit.Matern.aniso(grd, y, NU = 1)
}
\arguments{
\item{grd}{a two columns matrix with the spatial locations of the data}

\item{y}{the value of the response variable at the locations in grd}

\item{NU}{a number > 0, the smoothness parameter}
}
\value{
an optim object with the parameter estimates of the anisotropic Matern
}
\description{
Wrapper function for fitting an anisotropic Matern GP
}
