% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Matern.logLik.aniso.R
\name{fit.Matern.reml}
\alias{fit.Matern.reml}
\title{Wrapper function for fitting an isotropic Matern GP with REML}
\usage{
fit.Matern.reml(grd, y, Z, NU = 1)
}
\arguments{
\item{grd}{a two columns matrix with the spatial locations of the data}

\item{y}{the value of the response variable at the locations in grd}

\item{Z}{covariates as columns in a matrix}

\item{NU}{a number > 0, the smoothness parameter}
}
\value{
an optim object with the parameter estimates of the isotropic Matern
}
\description{
Wrapper function for fitting an isotropic Matern GP with REML
}
