% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.ns.R
\name{fivepoint.logLik}
\alias{fivepoint.logLik}
\title{Log likelihood function using a GMRF precision matrix}
\usage{
fivepoint.logLik(pars = c(log(0.1), 1, 0, log(0.1), log(1)), x, y, Z, symm)
}
\arguments{
\item{pars}{parameters defining the precision matrix; the range}

\item{x}{x-coordinates of grid}

\item{y}{y-coordinates of grid}

\item{Z}{response variable of data at locations}

\item{symm}{if TRUE, defines H using the symmetric square root, otherwise uses the Cholesky}
}
\value{
the negative log likelihood
}
\description{
Log likelihood function using a GMRF precision matrix
}
