% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.ns.R
\name{fivepointB.J}
\alias{fivepointB.J}
\title{Encode nonstationary GMRF precision matrix}
\usage{
fivepointB.J(x, y, grd, symm = FALSE)
}
\arguments{
\item{x}{x-coordinates of grid}

\item{y}{y-coordinates of grid}

\item{grd}{two column location of grid}

\item{symm}{if TRUE, defines H using the symmetric square root, otherwise uses the Cholesky}
}
\value{
a precision matrix
}
\description{
Encode nonstationary GMRF precision matrix
}
