% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Matern.logLik.aniso.R
\name{ilogit}
\alias{ilogit}
\title{Inverse logit / expit function}
\usage{
ilogit(x)
}
\arguments{
\item{x}{any number (log odds)}
}
\value{
the probability between 0 and 1
}
\description{
Inverse logit / expit function
}
