% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Matern.NS.Cov.R
\name{Isotropic.Matern.Exp.RQ.Nonstationary.Covs}
\alias{Isotropic.Matern.Exp.RQ.Nonstationary.Covs}
\title{Define a nonstationary covariance matrix based on a generalization of the
Matern with geometric anisotropy}
\usage{
Isotropic.Matern.Exp.RQ.Nonstationary.Covs(
  x,
  y,
  theta,
  nu,
  sigma = 1,
  family = "matern",
  verbose = FALSE
)
}
\arguments{
\item{x}{x-coordinate of locations on a regular rectangular grid}

\item{y}{y-coordinate of locations on a regular rectangular grid}

\item{theta}{correlation range parameter}

\item{nu}{smoothness parameter}

\item{sigma}{standard deviation parameter}

\item{family}{Matern, exp, or RQ (rational quadratic)}

\item{verbose}{whether to print progress or not}
}
\value{
a covariance matrix
}
\description{
Encode (local) spatially varying parameters (range theta, smoothness nu,
variance sigma) into a global nonstationarity covariance with local isotropy.
See Paciorek (2003) and Stein (2005).
}
