% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Matern.logLik.aniso.R
\name{Matern.logLik.aniso}
\alias{Matern.logLik.aniso}
\title{Log likelihood function for anisotropic Matern GP}
\usage{
Matern.logLik.aniso(
  pars = c(log(1.5), log(0.1), log(0.01), log(0.5), 0),
  grd,
  y,
  nu = 1
)
}
\arguments{
\item{pars}{covariance parameters: range in x, Matern variance, white noise
variance, range in y, and rotation angle}

\item{grd}{a two columns matrix with the spatial locations of the data}

\item{y}{the value of the response variable at the locations in grd}

\item{nu}{a number > 0, the smoothness parameter}
}
\value{
the negative log likelihood
}
\description{
Log likelihood function for anisotropic Matern GP
}
