% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Matern.logLik.aniso.R
\name{Matern.reml.aniso.loglik}
\alias{Matern.reml.aniso.loglik}
\title{Residual log likelihood function for anisotropic Matern GP}
\usage{
Matern.reml.aniso.loglik(
  pars = c(log(1.5), log(3), log(0.3), log(1.5), 0),
  grd,
  y,
  Z,
  nu = 1
)
}
\arguments{
\item{pars}{covariance parameters: range in x, Matern variance, white noise
variance, range in y, and rotation angle}

\item{grd}{a two columns matrix with the spatial locations of the data}

\item{y}{the value of the response variable at the locations in grd}

\item{Z}{covariates as columns in a matrix}

\item{nu}{a number > 0, the smoothness parameter}
}
\value{
the negative log likelihood
}
\description{
Residual log likelihood function for anisotropic Matern GP
}
