% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Matern.logLik.aniso.R
\name{Matern.reml.loglik}
\alias{Matern.reml.loglik}
\title{Residual log likelihood function for isotropic Matern GP}
\usage{
Matern.reml.loglik(pars = c(log(1.5), log(3), log(0.3)), grd, y, Z, nu = 1)
}
\arguments{
\item{pars}{covariance parameters: range parameter, Matern variance, and
white noise variance}

\item{grd}{a two columns matrix with the spatial locations of the data}

\item{y}{the value of the response variable at the locations in grd}

\item{Z}{covariates as columns in a matrix}

\item{nu}{a number > 0, the smoothness parameter}
}
\value{
the negative log likelihood
}
\description{
Residual log likelihood function for isotropic Matern GP
}
