% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nsMaternEstimates}
\alias{nsMaternEstimates}
\title{Local Matern covariance parameter estimates}
\format{
Local parameter estimates: thx, thy, and ang defining the local
geometric anisotropy, s2 and t2 defining the process and white noise
variances. Estimated on a grd (two columns matrix of locations)
}
\usage{
nsMaternEstimates
}
\description{
A dataset containing 30 simulations of a climate model ensemble
}
\keyword{datasets}
